PMBackend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
mysql-connector-python==9.2.0
